refactor: 레이어드 아키텍처로 클래스 분리
