+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -12,6 +12,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.dto.ReservationRequestDto;
+import roomescape.domain.entity.Reservation;
+import roomescape.domain.entity.ReservationTime;
 
 @RestController
 public class ReservationController {
