+++ src/main/java/roomescape/domain/dto/ReservationTimeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain.dto;
 
 public record ReservationTimeRequestDto(String startAt) {
 }
