+++ src/main/java/roomescape/domain/entity/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain.entity;
 
 public record Reservation(Long id, String name, String date, ReservationTime time) {
 }
