+++ src/main/java/roomescape/domain/entity/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain.entity;
 
 public record ReservationTime(Long id, String startAt) {
 }
